# PERF: Speed-up common case of loadtxt()ing non-hex floats.