@anntzer 2021-08-03
I added some more logic in the except branch to handle these cases and restore the original error message; because they are normally not exercised, the performance speedup remains present:
```
       before           after         ratio
     [f25905b5]       [f3f2e72a]
     <main>           <loadtxtfloat>
+     1.32±0.01μs      1.47±0.02μs     1.12  bench_io.Copy.time_memcpy('int16')
-      33.1±0.4ms       31.5±0.2ms     0.95  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('float32', 10000)
-         345±2ms          328±3ms     0.95  bench_io.LoadtxtCSVComments.time_comment_loadtxt_csv(100000)
-       354±0.7μs          336±3μs     0.95  bench_io.LoadtxtCSVComments.time_comment_loadtxt_csv(100)
-      9.66±0.2ms      9.16±0.05ms     0.95  bench_io.LoadtxtUseColsCSV.time_loadtxt_usecols_csv(2)
-      57.6±0.3μs       54.5±0.3μs     0.95  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('float32', 10)
-      33.4±0.2ms       31.6±0.1ms     0.95  bench_io.LoadtxtCSVComments.time_comment_loadtxt_csv(10000)
-         346±2ms          327±2ms     0.94  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('float64', 100000)
-         427±1ms        403±0.7ms     0.94  bench_io.LoadtxtCSVSkipRows.time_skiprows_csv(0)
-      17.9±0.1ms       16.9±0.1ms     0.94  bench_io.LoadtxtReadUint64Integers.time_read_uint64_neg_values(10000)
-         428±2ms          402±1ms     0.94  bench_io.LoadtxtCSVSkipRows.time_skiprows_csv(500)
-      19.7±0.1ms       18.4±0.2ms     0.94  bench_io.LoadtxtUseColsCSV.time_loadtxt_usecols_csv([1, 3, 5, 7])
-     1.81±0.02ms      1.69±0.01ms     0.93  bench_io.LoadtxtReadUint64Integers.time_read_uint64(1000)
-      17.9±0.1ms       16.7±0.1ms     0.93  bench_io.LoadtxtReadUint64Integers.time_read_uint64(10000)
-         388±2ms          360±2ms     0.93  bench_io.LoadtxtCSVSkipRows.time_skiprows_csv(10000)
-        16.2±1μs      14.7±0.01μs     0.90  bench_io.Copy.time_strided_copy('complex64')
```
(not sure what's happening with bench_io.Copy though)